Move state enums to enums package
