+++ freqtrade/enums/__init__.py
@@ -1,3 +1,5 @@
 # flake8: noqa: F401
+from freqtrade.enums.runmode import NON_UTIL_MODES, OPTIMIZE_MODES, TRADING_MODES, RunMode
 from freqtrade.enums.selltype import SellType
 from freqtrade.enums.signaltype import SignalType
+from freqtrade.enums.state import State
