+++ tests/data/test_dataprovider.py
@@ -5,9 +5,9 @@ import pytest
 from pandas import DataFrame
 
 from freqtrade.data.dataprovider import DataProvider
+from freqtrade.enums import RunMode
 from freqtrade.exceptions import ExchangeError, OperationalException
 from freqtrade.plugins.pairlistmanager import PairListManager
-from freqtrade.state import RunMode
 from tests.conftest import get_patched_exchange
